refactor:(심화 요구사항)공통으로 사용하는 로직 모듈화
